WSDEV-860 fix two column
