+++ src/blocks/Listing/Listing.tsx
@@ -4,10 +4,9 @@ import ListingEvent from './components/ListingEvent'
 import ListingNews from './components/ListingNews'
 import ListingPeople from './components/ListingPeople'
 import ListingVideo from './components/ListingVideo'
-import ListingTwoColumns from './components/ListingTwoColumns'
 
 export interface ListingProps {
-    type: 'base' | 'event' | 'news' | 'people' | 'twoColumn' | 'icon'
+    type: 'base' | 'event' | 'news' | 'people' | 'icon'
     header: string
     noborder: boolean
     data: ListProps[]
@@ -21,6 +20,7 @@ export interface ListProps {
     description?: string
     date: string
     datetime: string
+    icon: string
 }
 
 export interface ListTwoColumnProps {
@@ -40,16 +40,15 @@ export interface TypeProps {
 const Listing: React.FC<ListingProps> = (props): JSX.Element => {
     const { type } = props
 
-    const cardTypes: TypeProps = {
+    const listTypes: TypeProps = {
         base: <ListingBase {...props} />,
         icon: <ListingBase hasIcon {...props} />,
         event: <ListingEvent {...props} />,
         news: <ListingNews {...props} />,
         people: <ListingPeople {...props} />,
         video: <ListingVideo {...props} />,
-        twoColumn: <ListingTwoColumns {...props} />,
     }
-    return cardTypes[type || 'base']
+    return listTypes[type || 'base']
 }
 
 export default Listing
